Simplify port live check in pmap_svc.c
There's some hack in pmap_getport that will cause a service
to be unregistered from the portmap list if we find the port
is no longer in use. Apart from being a gross hack, it is
also a rather inefficient hack. Since we now restrict pmap
emulation to IPv4, we know the address is always 0.0.0.0,
so no need to mess with uaddr strings.

(The bind_check code is a huge messy no-op anyway, since
all ports are added with bind_check = FALSE).

Signed-off-by: okir@suse.de
Signed-off-by: Steve Dickson <steved@redhat.com>