Fix for warm start
If you use rpcbind with the warm start functionality, it will load *all*
registrations from the warm start files, including those for rpcbind
and portmap. This is wrong, as that information may be stale - a user
may specifically edit the netconfig file and restart rpcbind to change
the transports it supports. In this case we want the registrations
to match the status quo, rather than the status before the restart.

This patch changes read_warmstart() to merge the existing rpcb/pmap
lists, which contain only the rpcbind/portmap entries, with the saved
start lists, but ignoring any rpcbind/portmap entries present in the
warm start files.

Signed-off-by: Olaf Kirch <okir@suse.de>
Signed-off-by: Steve Dickson <steved@redhat.com>