Properly identify local root user over ipv4/v6
When an application registers a service through an inet transport,
rpcbind will always treat the owner as "unknown".  This allows random
users to unregister such services, and replace them with their own -
man-in-the-middle attacks for services like ypserv are trivial.

This patch changes pmapproc_change to check whether the call originated
from a priviliged local port, and if that is the case, it identifies the
caller as "superuser".

This mimics the way the current Linux portmap behaves.

Signed-off-by: Olaf Kirch <okir@suse.de>
Signed-off-by: Steve Dickson <steved@redhat.com>