nd: set SO_REUSEADDR on NC_TPI_COTS listening sockets
I previously sent this patch to the libtirpc-devel list but got no
response. Resending with wider distribution...

If we don't set SO_REUSEADDR, then if there are any sockets on this port
in TIME_WAIT state when rpcbind is restarted then that will prevent the
bind() call from succeeding.

Details of the problem are here:

    https://bugzilla.redhat.com/show_bug.cgi?id=597356

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>