Make superuser the owner of registers from the kernel.
Registers from the kernel do not come in on AF_LOCAL
sockets so the exact owners of the service can not be determined.
But given the fact the loopback address is used and a privilege
port used, it can be assumed the owner of the service should
be the superuser.

Signed-off-by: Steve Dickson <steved@redhat.com>