Make is_loopback check more permissive
This patch relaxes the is_loopback() check to its original meaning;
i.e. verify that the caller is local. We no longer check whether
the source port is privileged, for a number of reasons.

1) The existing check did not allow *any* non-root program to register
   a services via UDP or TCP transport. It did however allow
   *any* registration when using the AF_LOCAL transport.

2) Unregistration of services is only possible if the caller has
   the same "user name", i.e.
   	"superuser"
		for root (when connecting through AF_LOCAL sockets,
		or when using pmap_set with a privileged port)
	numeric uid
		for non-root users when connecting through AF_LOCAL
		sockets
	"unknown"
		for all other users

This seems safe enough to allow the removal of the privileged port
check in is_localhost.

Signed-off-by: Olaf Kirch <okir@suse.de>
Signed-off-by: Steve Dickson <steved@redhat.com>